+++ libs/coin-modules/coin-aptos/src/types/assets.ts
@@ -2,7 +2,10 @@ import type { Asset } from "@ledgerhq/coin-framework/api/types";
 
 export type AptosAsset = Asset<AptosTokenInformation>;
 
-export type AptosTokenInformation = { standard: string; contractAddress: string };
+export type AptosTokenInformation = {
+  standard: string;
+  contractAddress: string;
+};
 
 export type AptosExtra = Record<string, unknown>;