+++ libs/evm-tools/CHANGELOG.md
@@ -1,12 +1,20 @@
 # @ledgerhq/evm-tools
 
-## 1.6.3-nightly.0
+## 1.6.3
 
 ### Patch Changes
 
 - Updated dependencies [[`107f35a`](https://github.com/LedgerHQ/ledger-live/commit/107f35a0650412716b088a3503b86435e6d9cf47)]:
-  - @ledgerhq/live-env@2.9.0-nightly.0
-  - @ledgerhq/cryptoassets-evm-signatures@13.5.7-nightly.0
+  - @ledgerhq/live-env@2.9.0
+  - @ledgerhq/cryptoassets-evm-signatures@13.5.7
+
+## 1.6.3-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`107f35a`](https://github.com/LedgerHQ/ledger-live/commit/107f35a0650412716b088a3503b86435e6d9cf47)]:
+  - @ledgerhq/live-env@2.9.0-next.0
+  - @ledgerhq/cryptoassets-evm-signatures@13.5.7-next.0
 
 ## 1.6.2